src/sevenrad_stills/operations/blur_gaussian_mps.py
"""
GPU-accelerated Gaussian blur using Metal Performance Shaders.

This implementation uses Apple's Metal Performance Shaders (MPS) framework,
which provides hand-optimized Metal Shading Language kernels for maximum
performance on macOS and iOS devices. MPSImageGaussianBlur IS a custom
Metal kernel - it's Apple's production-grade MSL code.

Performance characteristics:
- Uses native Metal textures and command buffers
- Zero Python/Swift bridging overhead for computation
- Hand-optimized for Apple Silicon and AMD GPUs
- Expected to match or exceed MLX performance
"""

from typing import Any

import numpy as np
from PIL import Image

from sevenrad_stills.operations.base import BaseImageOperation

# Import Metal frameworks via PyObjC
# Note: These are loaded at runtime, type hints are for documentation only
try:
    import Metal
    import MetalPerformanceShaders as MPS  # noqa: N817
except ImportError as e:
    msg = "Metal Performance Shaders requires macOS with Metal support"
    raise ImportError(msg) from e

# Constants
RGB_CHANNELS = 3
RGBA_CHANNELS = 4


def create_metal_texture(
    device: Any,  # MTLDevice - PyObjC doesn't provide type stubs  # noqa: ANN401
    array: np.ndarray,
    pixel_format: int = 70,  # MTLPixelFormatRGBA8Unorm
) -> Any:  # MTLTexture - PyObjC doesn't provide type stubs  # noqa: ANN401
    """
    Create a Metal texture from a NumPy array.

    Args:
        device: Metal device to create texture on
        array: NumPy array with shape (height, width, channels) in uint8 format
        pixel_format: Metal pixel format (70 = MTLPixelFormatRGBA8Unorm)

    Returns:
        Metal texture containing the array data

    """
    height, width, channels = array.shape

    # Convert to uint8 if needed
    if array.dtype != np.uint8:
        array = np.clip(array, 0, 255).astype(np.uint8)

    # Ensure we have 4 channels for RGBA format
    if channels == RGB_CHANNELS:
        # Add alpha channel (fully opaque)
        rgba_array = np.dstack(
            [array, np.full((height, width, 1), 255, dtype=np.uint8)]
        )
    else:
        rgba_array = array.astype(np.uint8)

    # Create texture descriptor (MTLPixelFormatRGBA8Unorm = 70)
    texture_desc = Metal.MTLTextureDescriptor.texture2DDescriptorWithPixelFormat_width_height_mipmapped_(  # noqa: E501
        pixel_format, width, height, False
    )
    # MTLTextureUsageShaderRead = 1, MTLTextureUsageShaderWrite = 2
    texture_desc.setUsage_(1 | 2)  # ShaderRead | ShaderWrite

    # Create texture
    texture = device.newTextureWithDescriptor_(texture_desc)

    # Metal expects data in row-major order, RGBA8Unorm uses 1 byte per channel
    bytes_per_row = width * RGBA_CHANNELS  # 4 bytes per pixel (RGBA)
    region = Metal.MTLRegionMake2D(0, 0, width, height)

    # Copy data to texture - ensure C-contiguous array
    rgba_contiguous = np.ascontiguousarray(rgba_array)
    texture.replaceRegion_mipmapLevel_withBytes_bytesPerRow_(
        region, 0, rgba_contiguous, bytes_per_row
    )

    return texture


def texture_to_numpy(
    texture: Any,  # MTLTexture - PyObjC doesn't provide type stubs  # noqa: ANN401
    original_channels: int,
    dtype: type[np.uint8] | type[np.float32] = np.uint8,
) -> np.ndarray:
    """
    Convert a Metal texture back to a NumPy array.

    Args:
        texture: Metal texture to convert (RGBA8Unorm format)
        original_channels: Number of channels in original image (3 for RGB, 4 for RGBA)
        dtype: Target NumPy dtype (default: uint8)

    Returns:
        NumPy array with shape (height, width, original_channels)

    """
    width = texture.width()
    height = texture.height()

    # Allocate buffer for RGBA data (uint8 format, 4 channels)
    bytes_per_row = width * RGBA_CHANNELS  # 4 bytes per pixel
    buffer = np.zeros((height, width, RGBA_CHANNELS), dtype=np.uint8)

    # Read texture data - PyObjC requires passing buffer directly
    region = Metal.MTLRegionMake2D(0, 0, width, height)
    texture.getBytes_bytesPerRow_fromRegion_mipmapLevel_(
        buffer, bytes_per_row, region, 0
    )

    # Extract original channels (drop alpha if RGB)
    if original_channels == RGB_CHANNELS:
        buffer = buffer[:, :, :RGB_CHANNELS]

    # Convert to target dtype if needed
    if dtype != np.uint8:
        buffer = buffer.astype(dtype)

    return buffer


class GaussianBlurMPSOperation(BaseImageOperation):
    """
    Apply Gaussian blur using Metal Performance Shaders.

    This operation uses Apple's MPSImageGaussianBlur, which is a hand-optimized
    Metal Shading Language kernel. It represents the peak performance achievable
    for Gaussian blur on Apple hardware without writing custom MSL code.

    The MPS framework is optimized for:
    - Apple Silicon (M1/M2/M3/M4) GPUs
    - AMD GPUs in Intel Macs
    - Batch operations and image processing pipelines
    """

    def __init__(self) -> None:
        """Initialize the MPS Gaussian blur operation."""
        super().__init__("blur_gaussian_mps")

        # Create Metal device and command queue (shared across all blurs)
        self.device = Metal.MTLCreateSystemDefaultDevice()
        if self.device is None:
            msg = "Metal is not supported on this device"
            raise RuntimeError(msg)

        self.command_queue = self.device.newCommandQueue()
        if self.command_queue is None:
            msg = "Failed to create Metal command queue"
            raise RuntimeError(msg)

    def validate_params(self, params: dict[str, Any]) -> None:
        """
        Validate parameters for the Gaussian blur operation.

        Args:
            params: A dictionary containing:
                - sigma (float): The standard deviation for the Gaussian kernel.
                  Must be non-negative. A sigma of 0 returns the original image.

        Raises:
            ValueError: If parameters are invalid.

        """
        if "sigma" not in params:
            msg = "Gaussian blur requires a 'sigma' parameter."
            raise ValueError(msg)

        sigma = params["sigma"]
        if not isinstance(sigma, (int, float)):
            msg = f"Sigma must be a number, got {type(sigma)}."
            raise ValueError(msg)
        if sigma < 0:
            msg = f"Sigma must be non-negative, got {sigma}."
            raise ValueError(msg)

    def apply(self, image: Image.Image, params: dict[str, Any]) -> Image.Image:
        """
        Apply MPS Gaussian blur to the image.

        Args:
            image: The input PIL Image.
            params: A dictionary with a 'sigma' key.

        Returns:
            The blurred PIL Image.

        """
        self.validate_params(params)
        sigma: float = params["sigma"]

        # If sigma is zero, no blur is applied
        if sigma == 0:
            return image.copy()

        # Convert PIL image to NumPy array
        img_array = np.array(image)
        original_dtype = img_array.dtype
        original_channels = img_array.shape[2] if img_array.ndim == 3 else 1  # noqa: PLR2004

        # Handle grayscale by converting to RGB
        if image.mode == "L":
            img_array = np.stack([img_array] * RGB_CHANNELS, axis=-1)
        elif image.mode == "RGBA":
            # Process RGB channels only, preserve alpha
            rgb = img_array[:, :, :RGB_CHANNELS]
            alpha = img_array[:, :, RGB_CHANNELS]

            # Create textures
            input_texture = create_metal_texture(self.device, rgb)
            output_texture = create_metal_texture(self.device, rgb)

            # Create and configure blur filter
            blur = MPS.MPSImageGaussianBlur.alloc().initWithDevice_sigma_(
                self.device, sigma
            )

            # Execute blur
            command_buffer = self.command_queue.commandBuffer()
            blur.encodeToCommandBuffer_sourceTexture_destinationTexture_(
                command_buffer, input_texture, output_texture
            )
            command_buffer.commit()
            command_buffer.waitUntilCompleted()

            # Convert back to NumPy
            blurred_rgb = texture_to_numpy(output_texture, RGB_CHANNELS, original_dtype)

            # Recombine with alpha
            result = np.dstack([blurred_rgb, alpha])
            return Image.fromarray(result)

        # For RGB images
        # Create Metal textures
        input_texture = create_metal_texture(self.device, img_array)
        output_texture = create_metal_texture(self.device, img_array)

        # Create Gaussian blur filter
        blur = MPS.MPSImageGaussianBlur.alloc().initWithDevice_sigma_(
            self.device, sigma
        )

        # Create command buffer and encode blur operation
        command_buffer = self.command_queue.commandBuffer()
        blur.encodeToCommandBuffer_sourceTexture_destinationTexture_(
            command_buffer, input_texture, output_texture
        )

        # Execute and wait for completion
        command_buffer.commit()
        command_buffer.waitUntilCompleted()

        # Convert result back to NumPy array
        result = texture_to_numpy(output_texture, original_channels, original_dtype)

        # Handle grayscale output
        if image.mode == "L":
            result = result[:, :, 0]  # Extract single channel

        return Image.fromarray(result)
